bugfix: interface function signatures

+++ src/IguildCouncil.sol
@@ -6,7 +6,7 @@ interface GuildCouncilI {
     function _guildVerdict(bool guildAgreement, uint48 proposalId)
         external
         returns(bool success);
-    function _callGuildsToVote(uint256[] calldata guildsId, uint48 proposalId)
+    function _callGuildsToVote(uint48[] calldata guildsId, uint48 proposalId)
        external
        returns(bool);
     function sendSilver(address sender, address receiver, uint48 guildId, uint256 silverAmount)
